Add files via upload
I am in search of wisdom, I am failing at this attempt so far but I want to persevere. The code comes from openube and it wouldn't run without bugs and I genuinely cant figure it out. I would like to expand the sphere so I can later change that gameobject to a different modelled shape. Can you wise people help? 

this is the original https://github.com/openube/Unity-Pulse-Sensor but it's not working, I'm am new to this but I have been following the instructions so I'm confused as to why it's not working. 
Cheers